AI_DB/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.23,<0.32
SQLAlchemy>=2.0,<2.1
psycopg2-binary>=2.9,<3.0
pydantic>=2.6,<3.0
python-dotenv>=1.0,<2.0
alembic>=1.13,<2.0
boto3>=1.34,<2.0
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0
structlog>=24.1,<25.0
apscheduler>=3.10,<4.0
jinja2>=3.1,<4.0
aiogram>=3.4,<4.0
openai>=1.35,<2.0
aiohttp>=3.9,<4.0
python-multipart>=0.0.6,<0.1
pymorphy3>=2.0,<3.0